limpando main
