+++ main.py
@@ -1,48 +1,7 @@
 import pygame
 import random
-#import pyAlphabetIndicators
-
-letras = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J',
-          'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T',
-          'U', 'V', 'X', 'W', 'Y', 'Z']
-cor_acerto = [144, 238, 144] #verde
-cor_erro = [220, 20, 60] #vermelho
-class Indicadores:
-    def __init__(self, posicao_inicial, espacamento, cor_inicial):
-        global letras
-        self.letras = letras  # Lista de letras (pode ser o alfabeto)
-        self.posicao_inicial = posicao_inicial  # Posição inicial (x, y) na tela
-        self.espacamento = espacamento  # Espaçamento entre as letras
-        self.cor_inicial = cor_inicial  # Cor inicial das letras
-
-        self.indicadores = []  # Lista de indicadores de letras (letra, cor)
-        self.fonte = pygame.font.Font(None, 24)
-
-        # Inicializa os indicadores com as letras e a cor inicial
-        for letra in self.letras:
-            self.indicadores.append((letra, self.cor_inicial))
-
-    def atualizar(self, chute, palavra):
-        global cor_acerto
-        global cor_erro
-        for i, (letra, cor) in enumerate(self.indicadores):
-            if letra == chute:
-                # Verifica se o chute é correto e muda a cor
-                if chute in palavra:
-                    self.indicadores[i] = (letra, cor_acerto)
-                else:
-                    self.indicadores[i] = (letra, cor_erro)
-
-    def desenhar(self, tela):
-        x, y = self.posicao_inicial
-
-        for letra, cor in self.indicadores:
-            letra_surface = self.fonte.render(letra, True, cor)
-            letra_rect = letra_surface.get_rect()
-            letra_rect.topleft = (x, y)
-            tela.blit(letra_surface, letra_rect)
-            x += self.espacamento
-
+from indicadores import Indicadores
+from sprite import SpriteAnimation
 
 # Inicializando o pygame e criando a janela
 pygame.init()
@@ -63,20 +22,6 @@ for linha in range(0, spritesheet.get_height(), 400):#percorrendo a largura da s
 #Frames 0 a 5: inicializando
 #Frames 6 e 7, 8 e 9, 10 e 11, 12 e 13, 14 e 15, 16 e 17, por fim 18 e 19 representam as vidas sendo perdidas
 #Frame 20 é a derrota.
-
-#Classe SpriteAnimation que herda de pygame.sprite.Sprite
-class SpriteAnimation(pygame.sprite.Sprite):
-    def __init__(self, frames):
-        super().__init__()
-        self.frames = frames
-        self.frame_atual = 0
-        self.image = self.frames[self.frame_atual]
-        self.rect = self.image.get_rect()
-        self.rect.center = (TELA_LARGURA // 2, 200)
-
-    def update(self):
-        self.frame_atual = (self.frame_atual + 1) % len(self.frames)
-        self.image = self.frames[self.frame_atual]
 
 # numeros para verde piscina em RGB
 verdepis = [60, 85, 80]
